Move kvArrayToObject to ajax/user
